Database connection refactor
