+++ graphics/app.py
@@ -4,7 +4,7 @@ from flask import Flask, g
 from views import blueprint, Resources, Graphics, DisplayGraphics
 from flask.ext.restful import Api
 from client import Client
-from utils import db
+from utils.database import db
 
 def create_app(blueprint_only=False):
   app = Flask(__name__, static_folder=None)
