+++ graphics/config.py
@@ -2,6 +2,7 @@ SECRET_KEY = 'this should be changed'
 SQLALCHEMY_DATABASE_URI = ''
 IMAGE_PATH = ''
 INCLUDE_ARXIV = False
+SQLALCHEMY_BINDS = {}
 #This section configures this application to act as a client, for example to query solr via adsws
 CLIENT = {
   'TOKEN': 'we will provide an api key token for this application'
