+++ graphics/utils/graphics.py
@@ -15,8 +15,7 @@ from sqlalchemy import Column, Integer, String, DateTime, Boolean
 from sqlalchemy.ext.declarative import DeclarativeMeta
 from sqlalchemy.dialects import postgresql
 from flask.ext.sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
+from database import db, AlchemyEncoder, Graphics
 
 thumb_link = '<a href="%s" target="_new" border=0><img src="%s" width="100px"></a>'
 graph_link = '<a href="graphics" border=0><img src="%s"></a>'
@@ -24,34 +23,6 @@ ADSASS_img = '<img src="%s">'
 ADSASS_thmb_img  = '<img src="%s" width="100px">'
 ADSASS_thmb_link = '<a href="graphics" border=0>%s</a>'
 ADS_image_url = 'http://articles.adsabs.harvard.edu/cgi-bin/nph-iarticle_query?bibcode=%s&db_key=AST&page_ind=%s&data_type=GIF&type=SCREEN_VIEW'
-
-class AlchemyEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj.__class__, DeclarativeMeta):
-            # an SQLAlchemy class
-            fields = {}
-            for field in [x for x in dir(obj) if not x.startswith('_') and x != 'metadata']:
-                data = obj.__getattribute__(field)
-                try:
-                    json.dumps(data) # this will fail on non-encodable values, like other classes
-                    fields[field] = data
-                except TypeError:
-                    fields[field] = None
-            # a json-encodable dict
-            return fields
-
-        return json.JSONEncoder.default(self, obj)
-
-class Graphics(db.Model):
-  __tablename__='graphics'
-  __bind_key__ = 'graphics_db'
-  id = Column(Integer,primary_key=True)
-  bibcode = Column(String,nullable=False,index=True)
-  doi = Column(String)
-  source = Column(String)
-  eprint = Column(Boolean)
-  figures = Column(postgresql.JSON)
-  modtime = Column(DateTime)
 
 def get_graphics(bibcode):
     session = db.session()
